// src/types/theme.type.ts
export type AlignmentType = 'start' | 'center' | 'end';
export type TextAlignmentType =
    | 'auto'
    | 'justify'
    | 'left'
    | 'center'
    | 'right';
export type JustifyContentType =
    | 'flex-start'
    | 'flex-end'
    | 'center'
    | 'space-between'
    | 'space-around'
    | 'space-evenly';
export type AlignItemsType = 'flex-start' | 'flex-end' | 'center';

//Typography
export type TypographyType =
    | 'display'
    | 'headline'
    | 'title'
    | 'subheading'
    | 'body'
    | 'caption'
    | 'button';

//Color
export type ColorType =
    | 'background'
    | 'surface'
    | 'primary'
    | 'secondary'
    | 'success'
    | 'warning'
    | 'danger'
    | 'text'
    | 'caption';

//Modal
export type ModalSizeType = 'default' | 'half' | 'full';